edits based on PR comments
